@@ -2,18 +2,13 @@ class ApplicationController < ActionController::Base
   protect_from_forgery
 
   before_filter :rename_params
-  before_filter :ensure_domain
-
-  APP_DOMAIN = 'www.valuethisnow.com'
-
-  def ensure_domain
-    if request.env['HTTP_HOST'] != APP_DOMAIN
-      # HTTP 301 is a "permanent" redirect
-      redirect_to "https://#{APP_DOMAIN}", :status => 301
-    end
-  end
+  before_filter :configure_permitted_parameters, if: :devise_controller?
 
   rescue_from CanCan::AccessDenied do |exception|
+    redirect_to admin_dashboard_path, :alert => exception.message
+  end
+
+  def access_denied(exception)
     redirect_to admin_dashboard_path, :alert => exception.message
   end
 
@@ -23,6 +18,7 @@ class ApplicationController < ActionController::Base
 
   def after_sign_in_path_for(resource_or_scope)
     return session[:user_return_to] if session[:user_return_to]
+    
     if (current_user.admin?)
       admin_dashboard_path
     elsif(current_user.is_appraiser?)
@@ -33,10 +29,10 @@ class ApplicationController < ActionController::Base
   end
 
   def authenticate_admin_user!
-    authenticate_user! 
+    authenticate_user!
     unless current_user.admin?
       flash[:alert] = "This area is restricted to administrators only."
-      redirect_to root_path 
+      redirect_to root_path
     end
   end
 
@@ -60,6 +56,21 @@ class ApplicationController < ActionController::Base
   def check_admin
     if current_user.nil? || !current_user.admin?
       redirect_to root_path
+    end
+  end
+
+  def configure_permitted_parameters
+    devise_parameter_sanitizer.for(:sign_up) { |u| u.permit(:name, :email, :password, :password_confirmation) }
+  end
+
+  def ensure_signup_complete
+    # Ensure we don't go into an infinite loop
+    return if action_name == 'finish_signup'
+
+    # Redirect to the 'finish_signup' page if the user
+    # email hasn't been verified yet
+    if current_user && !current_user.email_verified?
+      redirect_to finish_signup_path(current_user)
     end
   end
 
