@@ -7,29 +7,34 @@ class DashboardController < ApplicationController
 		elsif @user.is_appraiser? and !@user.is_confirmed?
 			redirect_to appraiser_steps_path
 		else
-			
+
 			case @detail = params[:detail]
 				when "processing"
-					@appraisals = Appraisal.where("assigned_to = ? and status = ?",current_user,EActivityValueClaimed)
+					@appraisals = Appraisal.where("assigned_to = ? and status = ?",current_user,EActivityValueClaimed).order("updated_at desc")
 				when "unclaimed"
-					@appraisals = Appraisal.where('status = ? and (appraiser_referral is NULL or appraiser_referral =?)', EActivityValuePayed, current_user.referral_id)
+          @appraisals = Appraisal.where('(status = ? AND assigned_to IS null) OR (status = ? AND assigned_to = ?)', EActivityValuePayed, EActivityValuePayed, current_user.id)
           if Setting.is_set("show_all_appraisals","false")
-            @specializedAppraisals = Appraisal.select("appraisals.id").joins(:classification => {:category => {:skills => :appraiser}}).where('appraisals.status in (?) and categories.id in (?) and appraisals.appraiser_referral in (?)', [EActivityValuePayed, EActivityValueFinalized,EActivityValueClaimed ],current_user.skills.pluck(:category_id).uniq, [nil, current_user.referral_id]).pluck('appraisals.id').uniq
-            @appraisals = @appraisals.where("id in (?)",@specializedAppraisals)
+            @specializedAppraisals = Appraisal.select("appraisals.id").joins(:classification => {:category => {:skills => :appraiser}}).where('appraisals.status in (?) and categories.id in (?)', [EActivityValuePayed, EActivityValueFinalized,EActivityValueClaimed ],current_user.skills.pluck(:category_id).uniq).pluck('appraisals.id').uniq
+            @appraisals = @appraisals.where("id in (?)",@specializedAppraisals).order("updated_at desc")
           end
 				when "referrals"
-					@appraisals = Appraisal.where("appraiser_referral = ? and status = ?",current_user.referral_id,EActivityValuePayed)
+					@appraisals = Appraisal.where("appraiser_referral = ? and status = ?",current_user.referral_id,EActivityValuePayed).order("updated_at desc")
 				when "completed"
-					@appraisals = Appraisal.where("assigned_to = ? and status = ?",current_user,EActivityValueFinalized)
+					@appraisals = Appraisal.where("assigned_to = ? and status = ?",current_user,EActivityValueFinalized).order("updated_at desc")
         else
-					@appraisals = Appraisal.where('status = ?', EActivityValuePayed)
+          # For unclaimed status
+          @appraisals = Appraisal.where('(status = ? AND assigned_to IS null) OR (status = ? AND assigned_to = ?)', EActivityValuePayed, EActivityValuePayed, current_user.id).order("updated_at desc")
+
+          # For all status
+					# @appraisals = Appraisal.where("(status = ? AND assigned_to IS null) OR (assigned_to = ? )", EActivityValuePayed, current_user.id)
+          
           if Setting.is_set("show_all_appraisals","false")
             @specializedAppraisals = Appraisal.select("appraisals.id").joins(:classification => {:category => {:skills => :appraiser}}).where('appraisals.status in (?) and categories.id in (?)', [EActivityValuePayed, EActivityValueFinalized,EActivityValueClaimed ],current_user.skills.pluck(:category_id).uniq).pluck('appraisals.id').uniq
             @appraisals = @appraisals.where("id in (?)",@specializedAppraisals)
           end
 
 				end
-			@appraisals = @appraisals.visible	unless @appraisals.nil?			
+			@appraisals = @appraisals.visible	unless @appraisals.nil?
 
 			respond_to do |format|
 				format.html # index.html.haml
