@@ -1,12 +1,40 @@
 class OmniauthCallbacksController < Devise::OmniauthCallbacksController
-  def facebook
-    @user = User.find_for_facebook_oauth(request.env["omniauth.auth"], current_user)
+  # def facebook
+  #   @user = User.find_for_facebook_oauth(request.env["omniauth.auth"], current_user)
 
-    if @user.persisted?
-      sign_in_and_redirect @user, :event => :authentication
+  #   if @user.persisted?
+  #     sign_in_and_redirect @user, :event => :authentication
+  #   else
+  #     session["devise.facebook_data"] = env["omniauth.auth"]
+  #     redirect_to new_user_registration_url
+  #   end
+  # end
+
+  def self.provides_callback_for(provider)
+    class_eval %Q{
+      def #{provider}
+        @user = User.find_for_oauth(env["omniauth.auth"], current_user)
+
+        if @user.persisted?
+          sign_in_and_redirect @user, event: :authentication
+          set_flash_message(:notice, :success, kind: "#{provider}".capitalize) if is_navigational_format?
+        else
+          session["devise.#{provider}_data"] = env["omniauth.auth"]
+          redirect_to new_user_registration_url
+        end
+      end
+    }
+  end
+
+  [:twitter, :facebook, :linkedin].each do |provider|
+    provides_callback_for provider
+  end
+
+  def after_sign_in_path_for(resource)
+    if resource.email_verified?
+      super resource
     else
-      session["devise.facebook_data"] = env["omniauth.auth"]
-      redirect_to new_user_registration_url
+      finish_signup_path(resource)
     end
   end
 end
