@@ -2,7 +2,7 @@ class RegistrationsController < Devise::RegistrationsController
 
   def create
     # TODO This should be in the model
-    params["user"]["status"] = params["user"]["role"] == "customer" ? EAUserStatusConfirmed : EAUserStatusPending 
+    params["user"]["status"] = params["user"]["role"] == "customer" ? EAUserStatusConfirmed : EAUserStatusPending
     super
   end
 
@@ -13,6 +13,5 @@ class RegistrationsController < Devise::RegistrationsController
       after_sign_in_path_for(resource)
     end
   end
-
 
 end
