@@ -15,9 +15,9 @@ ActiveAdmin.setup do |config|
   # Set an optional image to be displayed for the header
   # instead of a string (overrides :site_title)
   #
-  # Note: Recommended image height is 21px to properly fit in the header
+  # Note: Aim for an image that's 21px high so it fits in the header.
   #
-  # config.site_title_image = "/images/logo.png"
+  # config.site_title_image = "logo.png"
 
   # == Default Namespace
   #
@@ -57,6 +57,23 @@ ActiveAdmin.setup do |config|
   # within the controller.
   config.authentication_method = :authenticate_admin_user!
 
+  # == User Authorization
+  #
+  # Active Admin will automatically call an authorization
+  # method in a before filter of all controller actions to
+  # ensure that there is a user with proper rights. You can use
+  # CanCanAdapter or make your own. Please refer to documentation.
+  config.authorization_adapter = ActiveAdmin::CanCanAdapter
+
+  # You can customize your CanCan Ability class name here.
+  # config.cancan_ability_class = "Ability"
+
+  # You can specify a method to be called on unauthorized access.
+  # This is necessary in order to prevent a redirect loop which happens
+  # because, by default, user gets redirected to Dashboard. If user
+  # doesn't have access to Dashboard, he'll end up in a redirect loop.
+  # Method provided here should be defined in application_controller.rb.
+  config.on_unauthorized_access = :access_denied
 
   # == Current User
   #
@@ -87,29 +104,55 @@ ActiveAdmin.setup do |config|
   config.logout_link_method = :delete
 
 
-  # == Admin Comments
+  # == Root
   #
-  # Admin comments allow you to add comments to any model for admin use.
-  # Admin comments are enabled by default.
+  # Set the action to call for the root path. You can set different
+  # roots for each namespace.
   #
   # Default:
+  # config.root_to = 'dashboard#index'
+
+
+  # == Admin Comments
+  #
+  # This allows your users to comment on any resource registered with Active Admin.
+  #
+  # You can completely disable comments:
   config.allow_comments = false
   #
-  # You can turn them on and off for any given namespace by using a
-  # namespace config block.
+  # You can disable the menu item for the comments index page:
+  # config.show_comments_in_menu = false
   #
-  # Eg:
-  #   config.namespace :without_comments do |without_comments|
-  #     without_comments.allow_comments = false
-  #   end
+  # You can change the name under which comments are registered:
+  # config.comments_registration_name = 'AdminComment'
+
+
+  # == Batch Actions
+  #
+  # Enable and disable Batch Actions
+  #
+  config.batch_actions = true
 
 
   # == Controller Filters
   #
   # You can add before, after and around filters to all of your
-  # Active Admin resources from here.
+  # Active Admin resources and pages from here.
   #
   # config.before_filter :do_something_awesome
+  
+  
+  # == Setting a Favicon
+  #
+  # config.favicon = '/assets/favicon.ico'
+
+
+  # == Removing Breadcrumbs
+  #
+  # Breadcrumbs are enabled by default. You can customize them for individual
+  # resources or you can disable them globally from here.
+  #
+  # config.breadcrumb = false
 
 
   # == Register Stylesheets & Javascripts
@@ -119,24 +162,83 @@ ActiveAdmin.setup do |config|
   # and feel.
   #
   # To load a stylesheet:
-  #   config.register_stylesheet 'my_stylesheet.css'
+    config.register_stylesheet '/assets/customize_activeadmin.css'
+    # config.register_javascript '/assets/customize_activeadmin.js'
   #
   # You can provide an options hash for more control, which is passed along to stylesheet_link_tag():
   #   config.register_stylesheet 'my_print_stylesheet.css', :media => :print
   #
   # To load a javascript file:
   #   config.register_javascript 'my_javascript.js'
-  # == Root
-  # Set the action to call for the root path. You can set different
-  # roots for each namespace.
-  # Default:
-  # config.root_to = 'dashboard#index'
-   
-  # == Batch Actions
-  # Enable and disable Batch Actions
-  config.batch_actions = true
-   
+
+
   # == CSV options
-  # Set the CSV builder separator (default is ",")
-  # config.csv_column_separator = ','
+  #
+  # Set the CSV builder separator
+  # config.csv_options = { :col_sep => ';' }
+  #
+  # Force the use of quotes
+  # config.csv_options = { :force_quotes => true }
+
+
+  # == Menu System
+  #
+  # You can add a navigation menu to be used in your application, or configure a provided menu
+  #
+  # To change the default utility navigation to show a link to your website & a logout btn
+  #
+  #   config.namespace :admin do |admin|
+  #     admin.build_menu :utility_navigation do |menu|
+  #       menu.add label: "My Great Website", url: "http://www.mygreatwebsite.com", html_options: { target: :blank }
+  #       admin.add_logout_button_to_menu menu
+  #     end
+  #   end
+  #
+  # If you wanted to add a static menu item to the default menu provided:
+  #
+  #   config.namespace :admin do |admin|
+  #     admin.build_menu :default do |menu|
+  #       menu.add label: "My Great Website", url: "http://www.mygreatwebsite.com", html_options: { target: :blank }
+  #     end
+  #   end
+
+
+  # == Download Links
+  #
+  # You can disable download links on resource listing pages,
+  # or customize the formats shown per namespace/globally
+  #
+  # To disable/customize for the :admin namespace:
+  #
+  #   config.namespace :admin do |admin|
+  #
+  #     # Disable the links entirely
+  #     admin.download_links = false
+  #
+  #     # Only show XML & PDF options
+  #     admin.download_links = [:xml, :pdf]
+  #
+  #     # Enable/disable the links based on block
+  #     #   (for example, with cancan)
+  #     admin.download_links = proc { can?(:view_download_links) }
+  #
+  #   end
+
+
+  # == Pagination
+  #
+  # Pagination is enabled by default for all resources.
+  # You can control the default per page count for all resources here.
+  #
+  # config.default_per_page = 30
+
+
+  # == Filters
+  #
+  # By default the index screen includes a “Filters” sidebar on the right
+  # hand side with a filter for each attribute of the registered model.
+  # You can enable or disable them for all resources here.
+  #
+  # config.filters = true
+
 end
